Add vertical wrap support, invalid opcode message
When sprites are drawn outside of the screen, they will wrap-around
vertically.
Whenever a program contains an invalid instruction, the
emulator will now print the opcode when it tries to execute it.
The collision flag VF is now set immediately instead of ORing it with
the pixel unset boolean, which saves a grand total of 1 cycle every
time the DXYN instruction is executed.